+++ src/engine/physics/cinematic.cpp
@@ -1,4 +1,4 @@
-#include "engine/physic/cinematic.h"
+#include "engine/physics/cinematic.h"
 
 void gravity(std::pair<double, double>* vel, double* mass){
     vel->second += *mass * PHX_CST_G;
